Add function to determine computer's choice
The new function 'getComputerChoice()' chooses pseudo-randomly
between 'Rock', 'Paper' and 'Scissors'. This is used by the
computer opponent to determine their choice. The function works
by using 'Math.random()' to choose a number between 0 and 3.
It should be exclusive at 3 and inclusive at 0.
After we have the number it's inputted into
the Math.floor() function to force the number to be either 0, 1 or 2.

There is a typo in a comment for the function and the odds are actually
exactly equal to each other which contradicts what was said inside
the function's comment. These will be fixed in the next commit.