Fix incorrect comment about computer's choice odds
Fixes the issue raised by last commit of incorrectly commenting on the
odds of the computer choosing between the three options. Beforehand
it was claimed that 'Scissors' had slightly lower odds but this is not
the case. In fact they have equal odds and if Math.random() had '1'
(the other border) being inclusive the function would run into
problems with Math.random() would be equal to 1 (since 1*3==3
and Math.floor(3)==3 which is not the as scissors and if it was
added to scissors would actually skew the odds for scissors).