Create function game() to loop 5 rounds
Creates the function game() that loops 5 rounds to create a game of
five rounds. It saves the total outcome by checking the string
returned by playRound() to find either 'Draw!', 'Win!' or 'Lose!'.
This is done by using the split() method. Currently it has
console.log's for testing purposes but those will be reworked or
removed in future commits.