Add external css for styling
Makes an external css that styles the buttons and puts the items close
to the center of the page. "justify-content: center" was first applied
to the body but this made the look inconsistent since the results class
changes with time. This was overcome by using top margin and works
because the layout is so simple. This commit also makes the text
displayed consistent.